Refactor code in Matching Service
